requirements.txt
APScheduler==3.6.3
black==25.1.0
cachetools==4.2.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
filelock==3.18.0
flake8==7.3.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyTelegramBotAPI==4.27.0
pytest==8.4.1
python-dotenv==1.1.1
python-telegram-bot==13.15
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
tornado==6.1
tqdm==4.67.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
